+++ ptlib/src/ptlib/msos/svcproc.cxx
@@ -334,6 +334,7 @@
 #pragma comment(lib,"user32.lib")
 #pragma comment(lib,"comdlg32.lib")
 
+
 #ifdef __USE_STL__
 #include <fstream>
 #else
@@ -610,7 +611,7 @@ int PSystemLog::Buffer::sync()
     // Trace system sets the ios stream width as the last thing it does before
     // doing a flush, which gets us here. SO now we can get a PTRACE looking
     // exactly like a PSYSTEMLOG of appropriate level.
-    unsigned traceLevel = log->width() -1 + PSystemLog::Warning;
+    unsigned traceLevel = (int)log->width() -1 + PSystemLog::Warning;
     log->width(0);
     if (traceLevel >= PSystemLog::NumLogLevels)
       traceLevel = PSystemLog::NumLogLevels-1;
